+++ DTCD-ButtonPanel/src/Plugin.js
@@ -1,11 +1,22 @@
 import pluginMeta from './Plugin.Meta';
 import PluginComponent from './PluginComponent.vue';
 
-import { PanelPlugin, LogSystemAdapter, EventSystemAdapter } from './../../DTCD-SDK';
+import { 
+  PanelPlugin, 
+  LogSystemAdapter, 
+  EventSystemAdapter 
+} from './../../DTCD-SDK';
 
 export class VisualizationText extends PanelPlugin {
-  #title;
+
   #eventSystem;
+  #vueComponent;
+
+  #config = {
+    title: '',
+    titleLeft:'',
+    buttonColor: 'theme_blueSec',
+  };
 
   static getRegistrationMeta() {
     return pluginMeta;
@@ -27,23 +38,31 @@ export class VisualizationText extends PanelPlugin {
       render: h => h(PluginComponent),
     }).$mount(selector);
 
-    this.vueComponent = view.$children[0];
-    this.#title = 'Кнопка';
+    this.#vueComponent = view.$children[0];
+  }
+
+  setVueComponentPropValue(prop, value) {
+    const methodName = `set${prop.charAt(0).toUpperCase() + prop.slice(1)}`;
+    if (this.#vueComponent[methodName]) {
+      this.#vueComponent[methodName](value)
+    } else {
+      throw new Error(`В компоненте отсутствует метод ${methodName} для присвоения свойства ${prop}`)
+    }
   }
 
   setPluginConfig(config = {}) {
-    const { title } = config;
+    const configProps = Object.keys(this.#config);
 
-    if (typeof title !== 'undefined') {
-      this.#title = title;
-      this.vueComponent.setTitle(title);
+    for (const [prop, value] of Object.entries(config)) {
+      if (!configProps.includes(prop)) continue;
+      this.setVueComponentPropValue(prop, value)
+
+      this.#config[prop] = value;
     }
   }
 
   getPluginConfig() {
-    const config = {};
-    if (this.#title) config.title = this.#title;
-    return config;
+    return { ...this.#config };
   }
 
   setFormSettings(config) {
@@ -61,11 +80,42 @@ export class VisualizationText extends PanelPlugin {
           component: 'text',
           propName: 'title',
           attrs: {
-            label: 'Отображаемый текст',
+            label: 'Текст в кнопке',
             required: true,
           },
         },
+        {
+          component: 'text',
+          propName: 'titleLeft',
+          attrs: {
+            label: 'Текст слева от кнопки',
+          },
+        },
+        {
+          component: 'select',
+          propName: 'buttonColor',
+          attrs: {
+            label: 'Выбрать цвет кнопки',
+          },
+          options: () => [
+            { label: 'Зеленый', value: 'theme_green' },
+            { label: 'Синий', value: 'theme_blueSec' },
+            { label: 'Красный', value: 'theme_red' },
+            { label: 'Серый', value: 'theme_secondary' },
+            { label: 'Прозрачный', value: 'theme_alfa' },
+          ],
+        },
       ],
     };
+  }
+
+  getState() {
+    return this.getPluginConfig();
+  }
+
+  setState(newState) {
+    if (typeof newState !== 'object' ) return;
+
+    this.setPluginConfig(newState);
   }
 }