+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+
+### Added
+
+- plugin state API
+- additional settings for the panel
+- options of displaying the panel through the component settings
+- the adaptive design of the panel
+
+### Changed
+
+- the design of panel
+- logic of the config
+- the color of the Container is set in styles file
+
+### Removed
+
+- textLeft when creating the panel
+
 ## [0.3.0]
 
 ### Changed